--- src/main/java/FileCorruptedException.java
@@ -1,5 +1,0 @@
-public class FileCorruptedException extends Exception {
-    public FileCorruptedException() {
-        super();
-    }
-}
